# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Intellios API
  version: 1.0.0
  description: |
    The Intellios API provides programmatic access to agent blueprint management,
    governance policy enforcement, and registry operations.

    ## Authentication
    All endpoints require either:
    - **Session cookie** (browser-based, via NextAuth)
    - **API key** (Bearer token): `Authorization: Bearer ik_live_...`

    ## Base URL
    `https://your-instance.intellios.ai`
servers:
  - url: https://your-instance.intellios.ai
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key (ik_live_...)
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Template:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        riskTier: { type: string, nullable: true }
        source: { type: string, enum: [built-in, community] }
        rating: { type: number, nullable: true, minimum: 1, maximum: 5 }
        usageCount: { type: integer }
        author: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
...
paths:
...
  /api/templates:
    get:
      summary: List agent templates (built-in and community)
      operationId: listTemplates
      tags: [Templates]
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Search by name
        - in: query
          name: source
          schema: { type: string, enum: [built-in, community] }
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items: { $ref: "#/components/schemas/Template" }
...